chore: Promise.all로 병렬적 import
