+++ src/label/components/LabelList.js
@@ -1,0 +1,27 @@
+import { $, request } from '../../utils/index.js';
+import { getLabelItemTpl } from '../../tpl.js';
+
+class LabelList {
+  $target;
+  constructor({ selector = '.label-list', setList, subscribe }) {
+    this.$target = $(selector);
+    subscribe((state) => this.render(state.labelList));
+    request({ url: '/labels' }).then((res) => {
+      setList(res);
+    });
+  }
+  render(labelList) {
+    const labelListTpl = labelList.reduce(
+      (acc, curr) =>
+        (acc += getLabelItemTpl({
+          name: curr.name,
+          color: curr.color,
+          description: curr.description,
+        })),
+      '',
+    );
+    this.$target.innerHTML = labelListTpl;
+  }
+}
+
+export default LabelList;
